Implement loan status and loan type features with corresponding data models and adapters
- Added LoanStatus and LoanType models along with their data entities.
- Created LoanStatusGateway and LoanTypeGateway interfaces for data access.
- Developed LoanStatusAdapter and LoanTypeAdapter for R2DBC integration.
- Introduced LoanStatusUseCase and LoanTypeUseCase for business logic handling.
- Implemented LoanTypeHandler and LoanStatusHandler for API request handling.
- Updated LoanApplication to use BigDecimal for loan amounts.
- Enhanced RequestRouterRest to include routes for loan types.
- Added DTOs for LoanType and updated LoanApplicationDTO for validation.

+++ domain/model/src/main/java/co/com/pragma/model/loanapplication/LoanApplication.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,7 +14,7 @@ import lombok.Setter;
 @Builder(toBuilder = true)
 public class LoanApplication {
     Long idLoan;
-    Long loanAmount;
+    BigDecimal loanAmount;
     Long termInMonths;
     String documentNumber;
     Long loanType;
